Make UnauthorizedHandler overridable
Two related changes here:
- Change from always throwing WebApplicationException when a user fails
to authenticate to throwing the result of the buildException method on
the unauthorized handler. This matches the pattern used in the built-in
BasicCredentialAuthFilter and OAuthCredentialAuthFilter.

- Add a `withUnauthorizedHandler` method to the bundle to set your own
UnauthorizedHandler. This gives you the ability to change the response
and exception thrown when a user fails to authenticate. For example, you
can redirect to a login page if you're building a browser application.